Coregidos errores varios

+++ ws-app-model/src/main/java/es/udc/isd060/runfic/model/inscripcion/Inscripcion.java
@@ -23,11 +23,11 @@ public class Inscripcion {
     }
 
 
-    public int getIdInscripcion() {
+    public Long getIdInscripcion() {
         return idInscripcion;
     }
 
-    public int getIdCarrera() {
+    public Long getIdCarrera() {
         return idCarrera;
     }
